backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
